restructure cartesian test suite
