+++ tests/cartesian_tests/integration_tests/multi_feature_tests/stencil_definitions.py
