+++ tests/cartesian_tests/integration_tests/multi_feature_tests/test_dace_parsing.py
@@ -25,8 +25,7 @@ from gt4py.cartesian import gtscript
 from gt4py.cartesian.gtscript import PARALLEL, computation, interval
 from gt4py.cartesian.stencil_builder import StencilBuilder
 from gt4py.storage import utils as storage_utils
-
-from ..utils import OriginWrapper
+from tests.cartesian_tests.utils import OriginWrapper
 
 
 dace = pytest.importorskip("dace")
