+++ tests/cartesian_tests/integration_tests/single_feature_tests/test_call_interface.py
@@ -20,9 +20,8 @@ import pytest
 import gt4py.cartesian.gtscript as gtscript
 import gt4py.storage as gt_storage
 from gt4py.cartesian.gtscript import Field, K
-
-from ..definitions import ALL_BACKENDS, CPU_BACKENDS
-from ..utils import DimensionsWrapper, OriginWrapper
+from tests.cartesian_tests.definitions import ALL_BACKENDS, CPU_BACKENDS
+from tests.cartesian_tests.utils import DimensionsWrapper, OriginWrapper
 
 
 def base_stencil(
