+++ tests/cartesian_tests/integration_tests/single_feature_tests/test_stencil_object.py
@@ -22,9 +22,8 @@ import pytest
 from gt4py import storage as gt_storage
 from gt4py.cartesian import gtscript
 from gt4py.cartesian.gtscript import PARALLEL, Field, computation, interval
-
-from ..definitions import ALL_BACKENDS
-from ..utils import OriginWrapper
+from tests.cartesian_tests.definitions import ALL_BACKENDS
+from tests.cartesian_tests.utils import OriginWrapper
 
 
 @pytest.mark.parametrize("backend", ["numpy"])
